Added NoTypeInference flag to Json and Xml TP
Previously the TypeProvider for Json and XML
tried to infer the real datatype by looking
at input string data. Ie. a string value of
"123" would be inferred as an int.

The inference is sometimes not desirable as seen
here:
https://github.com/fsharp/FSharp.Data/issues/656

NoTypeInference is by default set to false to
keep the previous behavior intact. When setting
NoTypeInference to true the TypeProvider no longer
tries to infer the datatype from input string data.